[ FEAT ] :  Reservation and cancel reservation finished

+++ Api/src/main/java/com/premierleague/reservation/api/service/UserService.java
@@ -1,8 +1,10 @@
 package com.premierleague.reservation.api.service;
 
 import com.premierleague.reservation.api.dtos.UserDTO;
+import com.premierleague.reservation.api.exceptions.UnauthorizedException;
 import com.premierleague.reservation.api.mappers.UserMapper;
 import com.premierleague.reservation.api.models.User;
+import com.premierleague.reservation.api.models.enums.Role;
 import com.premierleague.reservation.api.repositories.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,6 +18,9 @@ public class UserService {
     private UserRepository userRepository;
     @Autowired
     private  UserMapper userMapper;
+
+    @Autowired
+    private AuthenticationService authenticationService;
 
     public UserDTO userProfile(String username) {
         Optional<User> user = Optional.ofNullable(userRepository.findUserByUsername(username)
@@ -39,6 +44,12 @@ public class UserService {
             throw new RuntimeException("User not found with id: " + id);
         }
 
+        // check if iam an admin
+        if(!authenticationService.getRole().equals("ADMIN")){
+            throw new UnauthorizedException("You are not authorized to perform this action");
+        }
+
+
         userRepository.deleteById(id);
     }
 
@@ -56,5 +67,35 @@ public class UserService {
             throw new RuntimeException("User not found with username: " + username);
         }
         return user.get();
+    }
+
+    public void approveUser(Long id) {
+        if (!userRepository.existsById(id)) {
+            throw new RuntimeException("User not found with id: " + id);
+        }
+
+        // check if iam an admin
+        if(!authenticationService.getRole().equals("ADMIN")){
+            throw new UnauthorizedException("You are not authorized to perform this action");
+        }
+
+        User user = userRepository.findById(id).get();
+
+        if (!user.isRequestedRole()) {
+            throw new RuntimeException("User did not request a role");
+        }
+
+        user.setRole(Role.valueOf("EFA_MANAGER"));
+        userRepository.save(user);
+    }
+
+    public boolean requestManager() {
+        String username = authenticationService.getUsername();
+        User user = userRepository.findUserByUsername(username).get();
+
+        user.setRequestedRole(true);
+        userRepository.save(user);
+
+        return true;
     }
 }
